% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crule-package.R
\docType{package}
\name{crule-package}
\alias{crule}
\alias{crule-package}
\title{crule: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/crule}
  \item Report bugs at \url{https://github.com/andrewallenbruce/crule/issues}
}

}
\author{
\strong{Maintainer}: Andrew Allen Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
