% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_factory.R
\name{cruler}
\alias{cruler}
\title{Save filter conditions}
\usage{
cruler(x = expression(), .env = NULL)
}
\arguments{
\item{x}{expression}

\item{.env}{environment}
}
\value{
cruler object
}
\description{
Save a series of filter conditions, and support
logical operating among conditions
}
\examples{
c1 <- cruler(cut == "Fair")

c2 <- cruler(x > 8)

!c1

c1 | c2

c1 & c2

}
