% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_factory.R
\name{gather_expr}
\alias{gather_expr}
\title{Gather atomic expressions}
\usage{
gather_expr(x)
}
\arguments{
\item{x}{expression}
}
\value{
\verb{<chr>} vector of expressions
}
\description{
Gather atomic expressions
}
\examples{
ex <- rlang::expr(a == 2 & b == 3 | !b & x + 2)

gather_expr(ex)

}
