% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_factory.R
\name{is_atomic_expr}
\alias{is_atomic_expr}
\title{Is expression atomic?}
\usage{
is_atomic_expr(x)
}
\arguments{
\item{x}{expression}
}
\value{
logical
}
\description{
Is expression atomic?
}
\examples{
is_atomic_expr(rlang::expr(x))

is_atomic_expr(rlang::expr(!x))

is_atomic_expr(rlang::expr(x + y))

is_atomic_expr(rlang::expr(x > 1))

is_atomic_expr(rlang::expr(!x + y))

is_atomic_expr(rlang::expr(x > 1 | y < 2))

}
