% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{pos_name_to_code}
\alias{pos_name_to_code}
\title{Place of Service Name to Code}
\usage{
pos_name_to_code(x)
}
\arguments{
\item{x}{\verb{<chr>} vector of \code{pos} names}
}
\value{
\verb{<chr>} vector of \code{pos} codes
}
\description{
Place of Service Name to Code
}
\examples{

pos_name_to_code(c(
"TELEHEALTH PROVIDED OTHER THAN IN PATIENT HOME",
"TELEHEALTH PROVIDED OTHER THAN IN PT HOME",
"TELEHEALTH PROVIDED OTHER THAN PT HOME"))

}
